+++ changelog/changelog.md
@@ -6,6 +6,10 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot)
 
+## [WIP] - 2024-07-28
+
+[commit](https://github.com/sbrstrkkdwmdr/sbrbot)
+
 ### Fixed
 
 - version list could exceed page count