+++ changelog/changelog.md
@@ -6,9 +6,9 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot)
 
-## [WIP] - 2024-07-28
+## [3.18.1] - 2024-07-28
 
-[commit](https://github.com/sbrstrkkdwmdr/sbrbot)
+[commit](https://github.com/sbrstrkkdwmdr/sbrbot/commit/9d34e74fe7cfea20a73c72addef560f0283c958a)
 
 ### Fixed